TextTrimmer v1
