--- NavfertyExcelAddIn/LimitTextLength/ITextTrimmer.cs
@@ -1,11 +1,0 @@
-﻿#nullable enable
-
-using Microsoft.Office.Interop.Excel;
-
-namespace NavfertyExcelAddIn.LimitTextLength
-{
-	public interface ITextTrimmer
-	{
-		void DisplayTrimTextUI(Range range);
-	}
-}
