+++ NavfertyExcelAddIn/Localization/RibbonLabels.ru-RU.resx
@@ -240,4 +240,7 @@
   <data name="Feedback" xml:space="preserve">
     <value>Обратная связь</value>
   </data>
+  <data name="TrimTextByLength" xml:space="preserve">
+    <value>Обрезать текст</value>
+  </data>
 </root>
