+++ NavfertyExcelAddIn/Registry.cs
@@ -6,7 +6,6 @@ using Navferty.Common;
 using NavfertyExcelAddIn.Commons;
 using NavfertyExcelAddIn.DataValidation;
 using NavfertyExcelAddIn.FindFormulaErrors;
-using NavfertyExcelAddIn.LimitTextLength;
 using NavfertyExcelAddIn.ParseNumerics;
 using NavfertyExcelAddIn.StringifyNumerics;
 using NavfertyExcelAddIn.Transliterate;
@@ -47,12 +46,7 @@ namespace NavfertyExcelAddIn
 				.EnableInterfaceInterceptors()
 				.InterceptedBy(typeof(ExceptionLogger));
 
-			builder.RegisterType<EmptySpaceTrimmer>()
-				.As<IEmptySpaceTrimmer>()
-				.EnableInterfaceInterceptors()
-				.InterceptedBy(typeof(ExceptionLogger));
-
-			builder.RegisterType<EmptySpaceTrimmer>()
+			builder.RegisterType<TextTrimmer>()
 				.As<ITextTrimmer>()
 				.EnableInterfaceInterceptors()
 				.InterceptedBy(typeof(ExceptionLogger));
