+++ NavfertyExcelAddIn/WorksheetCellsEditing/ITextTrimmer.cs
@@ -2,9 +2,10 @@
 
 namespace NavfertyExcelAddIn.WorksheetCellsEditing
 {
-	public interface IEmptySpaceTrimmer
+	public interface ITextTrimmer
 	{
 		void TrimExtraSpaces(Range range);
 		void RemoveAllSpaces(Range range);
+		void TrimTextByLengthUIDisplay(Range range);
 	}
 }
