+++ NavfertyExcelAddIn/WorksheetCellsEditing/TrimTextByLengthUI/frmTrimParams.resx
